Move generated code into `generated` directory
