+++ packages/@azure/keyvault/lib/secretsModels.ts
@@ -1,7 +1,5 @@
-import * as Models from "./models";
-
 import * as msRest from "@azure/ms-rest-js";
-import { SecretAttributes } from "./models";
+import { SecretAttributes } from "./generated/models";
 
 
 export { SecretAttributes };
